Removed all comments/track changes for sharing.
Original file name had suffix 042613. Last modified date is 26 April 2013, 6:07PM.

In order to provide a cleaner looking document for HS to review, I removed all the comments and accepted all the changes that had been in the track changes system. So this version is sort of a break point in the editing process. All the comments and questions present in the document were erased in this version. I figured at that point if they were important questions they would come up again and if not they would go away. This version comes back edited by both Hugh and Becky.

	modified:   Sierran Mixed Conifer Draft Description.docx