Updated draft diagram
Saved a copy of the OCFW diagram and modified it (greatly) to fit with the conditions and numbers for DFTO. DFTO should be able to be quickly modified to fit MHW now since they are almost identical. Did it this way since DFTO diagram was made in December and was untouched since then.

	modified:   DFTO_Draft_Diagram.pdf
	modified:   DFTO_Draft_Diagram.svg